Use full path for imports, fixup post route
