+++ app/app.js
@@ -1,8 +1,8 @@
 import Ember from 'ember';
 import Resolver from 'ember/resolver';
 import loadInitializers from 'ember/load-initializers';
-import config from './config/environment';
-import { appReady, appUnload } from './utils/metrics';
+import config from 'pixelhandler-blog/config/environment';
+import { appReady, appUnload } from 'pixelhandler-blog/utils/metrics';
 
 Ember.MODEL_FACTORY_INJECTIONS = true;
