+++ app/routes/about.js
@@ -1,6 +1,6 @@
 import Ember from 'ember';
-import ResetScroll from '../mixins/reset-scroll';
-import RenderUsingTimings from '../mixins/render-using-timings';
+import ResetScroll from 'pixelhandler-blog/mixins/reset-scroll';
+import RenderUsingTimings from 'pixelhandler-blog/mixins/render-using-timings';
 
 export default Ember.Route.extend(ResetScroll, RenderUsingTimings, {
   measurementName: 'about_view'
