+++ app/routes/admin/edit.js
@@ -1,5 +1,5 @@
 import Ember from 'ember';
-import ResetScroll from '../../mixins/reset-scroll';
+import ResetScroll from 'pixelhandler-blog/mixins/reset-scroll';
 
 export default Ember.Route.extend(ResetScroll, {
   resourceName: 'post',
