+++ app/routes/posts.js
@@ -1,4 +1,4 @@
 import Ember from 'ember';
-import ResetScroll from '../mixins/reset-scroll';
+import ResetScroll from 'pixelhandler-blog/mixins/reset-scroll';
 
 export default Ember.Route.extend(ResetScroll);
