+++ app/routes/posts/index.js
@@ -1,7 +1,7 @@
 import Ember from 'ember';
-import RecordChunksMixin from '../../mixins/record-chunks';
-import ResetScroll from '../../mixins/reset-scroll';
-import RenderUsingTimings from '../../mixins/render-using-timings';
+import RecordChunksMixin from 'pixelhandler-blog/mixins/record-chunks';
+import ResetScroll from 'pixelhandler-blog/mixins/reset-scroll';
+import RenderUsingTimings from 'pixelhandler-blog/mixins/render-using-timings';
 
 export default Ember.Route.extend(
     ResetScroll, RecordChunksMixin, RenderUsingTimings, {
