+++ app/templates/post/detail.hbs
@@ -22,4 +22,4 @@
   </div>
 {{/if}}
 
-{{render "posts/comments"}}
+{{render "post/comments"}}
