+++ app/templates/posts/index.hbs
@@ -3,7 +3,7 @@
   {{#each post in model}}
     <dt class="Blog-list-date u-greenTxt">{{format-date post.date}}</dt>
     <dd class="Blog-list-link u-cf">
-      {{!link-to 'post' post}}{{post.title}}{{!link-to}}
+      {{#link-to 'post' post}}{{post.title}}{{/link-to}}
     </dd>
   {{/each}}
 </dl>
