+++ app/utils/metrics.js
@@ -1,6 +1,6 @@
 /*jshint unused:false*/
 import Ember from 'ember';
-import config from '../config/environment';
+import config from 'pixelhandler-blog/config/environment';
 
 /*
   Facade for User Timing API for Web Peformance testing
