+++ app/views/admin/create.js
@@ -1,4 +1,4 @@
 import Ember from 'ember';
-import AdminInputs from '../../mixins/admin-inputs';
+import AdminInputs from 'pixelhandler-blog/mixins/admin-inputs';
 
 export default  Ember.View.extend(AdminInputs);
